Make the "post" param required in REST API comments queries
If this parameter is omitted, it's possible to query for all editorial
comments across all posts. This functionality isn't exposed anywhere in
the UI, and isn't intentional... just inherited from the REST comments
controller.

Since there isn't an index field available to use, querying for all
workflows comments on a site with substantial amounts of content is
extremely slow. This avoids these queries by making the "post" field
required.