Subtract workflow count from comment count instead of overwriting
In 0.4.5, after #195 was merged, a bug was introduced in an Altis
application. If an application previously filtered `get_comments_number`
(for example, to not count author replies when showing a comment count),
that filtering would be undone by the filter in this module.

Example: if you had this code in an Altis application using Bylines,

```php
function get_author_reply_count( int $post_id ) : int {
	$byline_users = Bylines\get_all_byline_authors( $post_id );

	$comment_args = [
		'count'          => true,
		'post_id'        => $post_id,
		'status'         => 'approve',
		'author__in'     => $byline_users,
	];

	return get_comments( $comment_args );
}
add_filter( 'get_comments_number', 'filter_get_comments_number', 10, 2 );
```
on a post with 4 comments, 3 of them from the author, then after priority
10 the filter value would be 1. But when the filter introduced in #195
fires at priority 11, the value would be incorrectly reset back to 4,
since the new get_comments query from workflows ignores the prior value
passed in to the filter.